+++ maps/map.cub
@@ -4,7 +4,7 @@
 1000000111111111
 1000000000001111
 1000000000001111
-10000S00000111111
+10000N00000111111
 100000000011111
 1000000000111111
 1111111111111111
